restructured code for side bar
