+++ Frontend/src/pages/Dashboard.jsx
@@ -1,19 +1,17 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import {
-  Menu, Plus, Sun, Wind, Thermometer, MapPin, Camera, LogOut,
-  Settings, User, Home, Book, Bell, Droplets, Search, ChevronDown,
-  Cloud, Umbrella, ArrowRight, AlertCircle
+  Plus, Sun, Wind, Thermometer, MapPin, Camera, Cloud, 
+  Umbrella, ArrowRight, AlertCircle, Droplets
 } from 'lucide-react';
+import Sidebar from '../components/Sidebar';
 import { auth } from '../firebase';
 import { signOut } from 'firebase/auth';
 import { useNavigate } from 'react-router-dom';
-// Remove Webcam import since it's now in AddPlant
 import {
   fetchWeatherByLocation,
   searchLocations,
   calculateSearchRelevance
 } from '../services/weatherApi';
-// Import AddPlant component
 import AddPlant from '../components/AddPlant';
 
 const Dashboard = () => {
@@ -25,11 +23,10 @@ const Dashboard = () => {
   const [showLocationInput, setShowLocationInput] = useState(true);
   const [searchResults, setSearchResults] = useState([]);
   const [isSearching, setIsSearching] = useState(false);
-  const [showAddPlant, setShowAddPlant] = useState(false); // Replace showCamera state
+  const [showAddPlant, setShowAddPlant] = useState(false); 
   const [plantName, setPlantName] = useState("");
   const [activeNavItem, setActiveNavItem] = useState('Home');
   const [isSearchFocused, setIsSearchFocused] = useState(false);
-  // Remove webcamRef since it's now in AddPlant
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -57,7 +54,7 @@ const Dashboard = () => {
   };
 
   const handleAddPlant = () => {
-    setShowAddPlant(true); // Update to use showAddPlant instead of showCamera
+    setShowAddPlant(true); 
   };
 
   // New function to handle the plant data from AddPlant
@@ -118,71 +115,12 @@ const Dashboard = () => {
 
   return (
     <div className="flex h-screen bg-gray-50">
-      {/* Sidebar */}
-      <div className={`bg-white h-full shadow-lg transition-all duration-300 ease-in-out ${
-        isMenuOpen ? 'w-64' : 'w-16'
-      } flex flex-col border-r border-gray-100`}>
-        {/* Hamburger Menu */}
-        <div className="p-4 flex justify-center">
-          <button 
-            onClick={() => setIsMenuOpen(!isMenuOpen)} 
-            className="p-2 rounded-lg hover:bg-gray-100 transition-colors"
-          >
-            <Menu size={22} className="text-gray-600" />
-          </button>
-        </div>
-
-        {/* Navigation Items - Only show when menu is open */}
-        <div className="flex-1 px-2">
-          {isMenuOpen && (
-            <div className="space-y-4 mt-4">
-              <NavItem 
-                icon={<Home size={20} />} 
-                label="Home" 
-                isActive={activeNavItem === 'Home'} 
-                onClick={() => setActiveNavItem('Home')}
-              />
-              <NavItem 
-                icon={<Book size={20} />} 
-                label="Garden Log" 
-                isActive={activeNavItem === 'Garden Log'} 
-                onClick={() => setActiveNavItem('Garden Log')}
-              />
-              <NavItem 
-                icon={<Bell size={20} />} 
-                label="Reminders" 
-                isActive={activeNavItem === 'Reminders'} 
-                onClick={() => setActiveNavItem('Reminders')}
-              />
-              <NavItem 
-                icon={<User size={20} />} 
-                label="Profile" 
-                isActive={activeNavItem === 'Profile'} 
-                onClick={() => setActiveNavItem('Profile')}
-              />
-              <NavItem 
-                icon={<Settings size={20} />} 
-                label="Settings" 
-                isActive={activeNavItem === 'Settings'} 
-                onClick={() => setActiveNavItem('Settings')}
-              />
-            </div>
-          )}
-        </div>
-
-        {/* Logout Button - Always visible */}
-        <div className="p-2 border-t border-gray-100">
-          <button 
-            onClick={handleLogout} 
-            className={`flex items-center justify-center w-full p-2 rounded-lg text-red-500 hover:bg-red-50 transition-colors ${
-              isMenuOpen ? 'justify-start px-4' : 'justify-center'
-            }`}
-          >
-            <LogOut size={20} />
-            {isMenuOpen && <span className="ml-3 font-medium">Log Out</span>}
-          </button>
-        </div>
-      </div>
+      <Sidebar 
+        isMenuOpen={isMenuOpen}
+        setIsMenuOpen={setIsMenuOpen}
+        activeNavItem={activeNavItem}
+        setActiveNavItem={setActiveNavItem}
+      />
 
       {/* Main Content */}
       <div className="flex-1 overflow-y-auto">
@@ -393,21 +331,6 @@ const Dashboard = () => {
     </div>
   );
 };
-
-// Update NavItem component to remove isOpen prop since it's controlled by parent visibility
-const NavItem = ({ icon, label, isActive, onClick }) => (
-  <div 
-    onClick={onClick}
-    className={`flex items-center space-x-4 p-2.5 rounded-lg cursor-pointer transition-colors ${
-      isActive 
-        ? 'bg-green-50 text-green-600' 
-        : 'hover:bg-gray-100 text-gray-600'
-    }`}
-  >
-    <div className={isActive ? 'text-green-600' : 'text-gray-500'}>{icon}</div>
-    <span className={`${isActive ? 'font-medium' : ''} transition-all`}>{label}</span>
-  </div>
-);
 
 const WeatherItem = ({ icon, label, value }) => (
   <div className="flex items-center p-3 bg-white rounded-lg shadow-sm hover:shadow hover:bg-gray-50 transition-all">
